+++ tests/audio/test_tags.py
@@ -120,7 +120,7 @@ class TagsToTrackTest(unittest.TestCase):
                       num_tracks=2, num_discs=3,
                       musicbrainz_id='albumid', artists=[albumartist])
 
-        self.track = Track(name='track',
+        self.track = Track(name='track', date='2006-01-01',
                            genre='genre', track_no=1, disc_no=2,
                            comment='comment', musicbrainz_id='trackid',
                            album=album, bitrate=1000, artists=[artist],
@@ -184,7 +184,9 @@ class TagsToTrackTest(unittest.TestCase):
     def test_missing_track_date(self):
         del self.tags['date']
         self.check(
-            self.track.replace(album=self.track.album.replace(date=None)))
+            self.track.replace(
+                album=self.track.album.replace(date=None)
+            ).replace(date=None))
 
     def test_multiple_track_date(self):
         self.tags['date'].append('2030-01-01')